Improving error reporting

+++ src/pipeline/mod.rs
@@ -4,8 +4,12 @@ use std::collections::HashMap;
 use std::fs;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
+use std::time::Duration;
 
 use anyhow::{anyhow, Context};
+use console::{style, Term};
+use indicatif::{ProgressBar, ProgressStyle};
+use serde::de::Error;
 use tera::Tera;
 
 use cdoc::config::OutputFormat;
@@ -30,6 +34,18 @@ pub struct Pipeline {
     base_tera: Tera,
     shortcode_tera: Tera,
     cached_contexts: HashMap<OutputFormat, GeneratorContext>,
+}
+
+
+pub fn print_err<T>(res: anyhow::Result<T>) -> Option<T> {
+    match res {
+        Ok(s) => Some(s),
+        Err(e) => {
+            eprintln!("{} {}", style("Error:").red().bold(), e);
+            e.chain().skip(1).for_each(|cause| eprintln!(" {} {}", style("caused by:").bold(), cause));
+            None
+        }
+    }
 }
 
 impl Pipeline {
@@ -77,8 +93,18 @@ impl Pipeline {
         }
     }
 
+    pub fn reload_shortcode_tera(&mut self) -> anyhow::Result<()> {
+        Ok(self.shortcode_tera.full_reload()?)
+    }
+
+    pub fn reload_base_tera(&mut self) -> anyhow::Result<()> {
+        Ok(self.base_tera.full_reload()?)
+    }
+
     pub fn build_single(&mut self, path: PathBuf) -> anyhow::Result<()> {
         println!("build single");
+        println!("{} {}", style("Building file").bold(), path.display());
+        println!("{}", style("-".repeat(60)).blue());
         let item = self.doc_from_path(path)?;
         let item2 = item.clone();
 
@@ -90,7 +116,9 @@ impl Pipeline {
         })?;
 
         for format in self.project_config.outputs.clone() {
-            let output = self.process_document(&loaded.doc, format)?;
+            print!("format: {}", style(format).bold());
+            let output = self.process_document(&loaded.doc, format);
+            let output = print_err(output).flatten();
 
             if let Some(output) = output {
                 let context = self
@@ -103,6 +131,10 @@ impl Pipeline {
 
                 self.get_generator(format)
                     .generate_single(output, item2.clone(), context)?;
+
+                println!(" {}", style("done").green());
+            } else {
+                println!(" {}", style("no output").yellow());
             }
         }
 
@@ -216,17 +248,25 @@ impl Pipeline {
         Ok(item)
     }
 
-    pub fn build_all(&mut self) -> Result<(), anyhow::Error> {
+    pub fn build_all(&mut self, remove_existing: bool) -> Result<(), anyhow::Error> {
         let build_path = self.project_path.join("build");
 
-        if build_path.exists() {
+        if remove_existing && build_path.exists() {
             fs::remove_dir_all(build_path)?;
         }
 
         let loaded = self.load_all()?;
 
+        println!("{}", style("=".repeat(60)).blue());
+        println!("{} ({} files)", style("Building project").bold(), loaded.len());
+        println!("{}", style("-".repeat(60)).blue());
+
+        let mut all_errs = Vec::new();
+
         for format in &self.project_config.outputs {
-            let output = self.process_all(loaded.clone(), *format)?;
+            println!("format: {}", style(format).bold());
+            let (output, mut errs) = self.process_all(loaded.clone(), *format);
+            all_errs.append(&mut errs);
             let context = GeneratorContext {
                 root: self.project_path.to_path_buf(),
                 project: output,
@@ -235,22 +275,48 @@ impl Pipeline {
             };
             self.cached_contexts.insert(*format, context.clone());
 
-            self.get_generator(*format)
+            // print!("[generating output");
+
+            let res = self.get_generator(*format)
                 .generate(context.clone())
-                .with_context(|| format!("Error while generating {}", format))?;
+                .with_context(|| format!("Could not generate {}", format));
 
-            self.cached_contexts.insert(*format, context);
+            match res {
+                Err(e) => all_errs.push(e),
+                Ok(_) => {
+                    self.cached_contexts.insert(*format, context);
 
-            if let Some(parser) = self.project_config.parsers.get(format) {
-                let move_ctx = MoveContext {
-                    project_path: self.project_path.to_path_buf(),
-                    build_dir: self.get_build_path(*format),
-                    settings: parser.settings.clone(),
-                };
+                    if let Some(parser) = self.project_config.parsers.get(format) {
+                        // print!(", copying additional files");
+                        let move_ctx = MoveContext {
+                            project_path: self.project_path.to_path_buf(),
+                            build_dir: self.get_build_path(*format),
+                            settings: parser.settings.clone(),
+                        };
 
-                Mover::traverse_dir(self.project_path.join("content").to_path_buf(), &move_ctx)?;
+                        Mover::traverse_dir(self.project_path.join("content").to_path_buf(), &move_ctx)?;
+                    }
+                }
             }
         }
+
+
+        println!("{}", style("-".repeat(60)).blue());
+        if all_errs.is_empty() {
+            println!("{}", style("Project built successfully").green().bold());
+        } else {
+            let len = all_errs.len();
+            all_errs.into_iter().for_each(|e| {
+                eprintln!("{} {}", style("Error:").red().bold(), e);
+                e.chain().skip(1).for_each(|cause| eprintln!(" {} {}", style("caused by:").bold(), cause));
+            });
+            println!("{}", style("-".repeat(60)).blue());
+
+            println!("{}", style(format!("Project built with non-critical errors ({} total)", len)).yellow().bold());
+        }
+        println!("{}", style("=".repeat(60)).blue());
+
+
         Ok(())
     }
 
@@ -276,16 +342,57 @@ impl Pipeline {
         &self,
         project: Project<String>,
         format: OutputFormat,
-    ) -> anyhow::Result<Project<Option<Document<RenderResult>>>> {
-        project
+    ) -> (Project<Option<Document<RenderResult>>>, Vec<anyhow::Error>) {
+        let spinner = ProgressStyle::with_template("{prefix:.bold.dim} {spinner} {wide_msg}").unwrap().tick_chars("⠁⠂⠄⡀⢀⠠⠐⠈ ");
+        let pb = ProgressBar::new(0);
+        pb.set_style(spinner);
+        // pb.set_prefix(format!("[{}/?]", i + 1));
+
+        let mut errs = Vec::new();
+
+        let res = project
             .into_iter()
             .map(|i| {
-                i.map_doc(|doc| {
-                    self.process_document(&doc, format)
-                        .with_context(|| format!("Failed to parse document {}", doc.path.display()))
-                })
+                pb.set_message(format!("{}", i.doc.path.display()));
+                pb.inc(1);
+
+                let res = self.process_document(&i.doc, format)
+                    .with_context(|| format!("Failed to process document – {}", style(format!("content/{}", i.doc.path.display())).italic()));
+
+                let res = match res {
+                    Ok(good) => good,
+                    Err(e) => {
+                        errs.push(e);
+                        None
+                    }
+                };
+                // let res = print_err(res);
+
+                ItemDescriptor {
+                    part_id: i.part_id,
+                    chapter_id: i.chapter_id,
+                    part_idx: i.part_idx,
+                    chapter_idx: i.chapter_idx,
+                    doc_idx: i.doc_idx,
+                    doc: ProjectItem {
+                        id: i.doc.id,
+                        format: i.doc.format,
+                        path: i.doc.path,
+                        content: Arc::new(res),
+                    },
+                    files: i.files,
+                }
             })
-            .collect()
+            .collect::<Project<Option<Document<RenderResult>>>>();
+
+        if errs.is_empty() {
+            pb.finish_with_message(format!("document processing {}", style("success").green()));
+        } else {
+            pb.finish_with_message(format!("document processing {}", style(format!("({} errors)", errs.len())).red()));
+        }
+        // pb.finish_with_message(format!("Done"));
+
+        (res, errs)
     }
 
     fn process_document(
@@ -294,7 +401,7 @@ impl Pipeline {
         format: OutputFormat,
     ) -> anyhow::Result<Option<Document<RenderResult>>> {
         let doc = item.format.loader().load(&item.content)?;
-
+        // println!("doc {}", item.path.display());
         if format.no_parse() {
             Ok(Some(Document {
                 content: "".to_string(),
@@ -313,10 +420,10 @@ impl Pipeline {
             let res = self
                 .project_config
                 .parsers
-                .get(&format)
-                .ok_or_else(|| anyhow!("No spec found"))?
+                .get(&format).ok_or_else(|| anyhow!("Invalid format"))?
                 .parse(&doc, &meta, &processor_ctx)?;
 
+            // let res = print_err(res)?;
             if let Some(renderer) = format.renderer() {
                 Ok(Some(renderer.render(&res)))
             } else {
